requirements.txt
annotated-types==0.7.0
certifi==2024.7.4
charset-normalizer==3.3.2
idna==3.7
mysqlclient==2.2.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
pydantic==2.8.2
pydantic_core==2.20.1
pyTelegramBotAPI==4.22.1
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
SQLAlchemy==2.0.32
telebot==0.0.5
typing_extensions==4.12.2
urllib3==2.2.2
pytest==7.4.0
pytest-cov==4.1.0